Start reorg.
