+++ politeiawww/user.go
@@ -11,7 +11,6 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"regexp"
 	"sort"
 	"strconv"
 	"strings"
@@ -26,39 +25,6 @@ import (
 	"golang.org/x/crypto/bcrypt"
 )
 
-const (
-	LoginAttemptsToLockUser = 5
-
-	// Number of attempts until totp locks until the next window
-	totpFailedAttempts = 2
-
-	// Route to reset password at GUI
-	ResetPasswordGuiRoute = "/password" // XXX what is this doing here?
-
-	emailRegex = `^[a-zA-Z0-9.!#$%&'*+/=?^_` +
-		"`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?" +
-		"(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$"
-)
-
-var (
-	validUsername = regexp.MustCompile(createUsernameRegex())
-	validEmail    = regexp.MustCompile(emailRegex)
-
-	// resetPasswordMinWaitTime is the minimum amount of time to wait
-	// before sending a response back to the client for the reset
-	// password route. This is done to prevent an attacker from being
-	// able to execute a timing attack to determine if the provided
-	// email address is the user's valid email address.
-	resetPasswordMinWaitTime = 500 * time.Millisecond
-
-	// loginMinWaitTime is the minimum amount of time to wait before
-	// the server sends a response to the client for the login route.
-	// This is done to prevent an attacker from being able to execute
-	// a timing attack to determine whether the ErrorStatusInvalidLogin
-	// response is specific to a bad email or a bad password.
-	loginMinWaitTime = 500 * time.Millisecond
-)
-
 // loginReply is used to pass the results of the login command between go
 // routines.
 type loginResult struct {
@@ -71,26 +37,6 @@ type loginResult struct {
 type resetPasswordResult struct {
 	reply www.ResetPasswordReply
 	err   error
-}
-
-// createUsernameRegex generates a regex based on the policy supplied valid
-// characters in a user name.
-func createUsernameRegex() string {
-	var buf bytes.Buffer
-	buf.WriteString("^[")
-
-	for _, supportedChar := range www.PolicyUsernameSupportedChars {
-		if len(supportedChar) > 1 {
-			buf.WriteString(supportedChar)
-		} else {
-			buf.WriteString(`\` + supportedChar)
-		}
-	}
-	buf.WriteString("]{")
-	buf.WriteString(strconv.Itoa(www.PolicyMinUsernameLength) + ",")
-	buf.WriteString(strconv.Itoa(www.PolicyMaxUsernameLength) + "}$")
-
-	return buf.String()
 }
 
 // convertWWWUserFromDatabaseUser converts a user User to a www User.
@@ -152,451 +98,6 @@ func filterUserPublicFields(user www.User) www.User {
 	}
 }
 
-// formatUsername normalizes a username to lowercase without leading and
-// trailing spaces.
-func formatUsername(username string) string {
-	return strings.ToLower(strings.TrimSpace(username))
-}
-
-// validateUsername verifies that a username adheres to required policy.
-func validateUsername(username string) error {
-	if username != formatUsername(username) {
-		log.Tracef("validateUsername: not normalized: %s %s",
-			username, formatUsername(username))
-		return www.UserError{
-			ErrorCode: www.ErrorStatusMalformedUsername,
-		}
-	}
-	if len(username) < www.PolicyMinUsernameLength ||
-		len(username) > www.PolicyMaxUsernameLength {
-		log.Tracef("validateUsername: not within bounds: %s",
-			username)
-		return www.UserError{
-			ErrorCode: www.ErrorStatusMalformedUsername,
-		}
-	}
-	if !validUsername.MatchString(username) {
-		log.Tracef("validateUsername: not valid: %s %s",
-			username, validUsername.String())
-		return www.UserError{
-			ErrorCode: www.ErrorStatusMalformedUsername,
-		}
-	}
-	return nil
-}
-
-// validatePassword verifies that a password adheres to required policy.
-func validatePassword(password string) error {
-	if len(password) < www.PolicyMinPasswordLength {
-		return www.UserError{
-			ErrorCode: www.ErrorStatusMalformedPassword,
-		}
-	}
-
-	return nil
-}
-
-// validatePubKey verifies that the provided public key is a valid ed25519
-// public key.
-func validatePubKey(publicKey string) error {
-	pk, err := hex.DecodeString(publicKey)
-	if err != nil {
-		log.Debugf("validatePubKey: decode hex string "+
-			"failed for '%v': %v", publicKey, err)
-		return www.UserError{
-			ErrorCode: www.ErrorStatusInvalidPublicKey,
-		}
-	}
-
-	var emptyPK [identity.PublicKeySize]byte
-	switch {
-	case len(pk) != len(emptyPK):
-		log.Debugf("validatePubKey: invalid size: %v",
-			publicKey)
-		return www.UserError{
-			ErrorCode: www.ErrorStatusInvalidPublicKey,
-		}
-	case bytes.Equal(pk, emptyPK[:]):
-		log.Debugf("validatePubKey: key is empty: %v",
-			publicKey)
-		return www.UserError{
-			ErrorCode: www.ErrorStatusInvalidPublicKey,
-		}
-	}
-
-	return nil
-}
-
-// validateSignature validates an incoming signature against the specified
-// public key and message. This function assumes the provided public key is
-// valid.
-func validateSignature(pubKey string, signature string, elements ...string) error {
-	sig, err := util.ConvertSignature(signature)
-	if err != nil {
-		return www.UserError{
-			ErrorCode: www.ErrorStatusInvalidSignature,
-		}
-	}
-	b, err := hex.DecodeString(pubKey)
-	if err != nil {
-		return www.UserError{
-			ErrorCode: www.ErrorStatusInvalidPublicKey,
-		}
-	}
-	pk, err := identity.PublicIdentityFromBytes(b)
-	if err != nil {
-		return err
-	}
-	var msg string
-	for _, v := range elements {
-		msg += v
-	}
-	if !pk.VerifyMessage([]byte(msg), sig) {
-		return www.UserError{
-			ErrorCode: www.ErrorStatusInvalidSignature,
-		}
-	}
-	return nil
-}
-
-// userIsLocked returns whether the user's account has been locked due to
-// failed login attempts.
-func userIsLocked(failedLoginAttempts uint64) bool {
-	return failedLoginAttempts >= LoginAttemptsToLockUser
-}
-
-// newVerificationTokenAndExpiry returns a byte slice of random data that is
-// the size of a verification token and a UNIX timestamp that represents the
-// expiration of the token.
-func newVerificationTokenAndExpiry() ([]byte, int64, error) {
-	tokenb, err := util.Random(www.VerificationTokenSize)
-	if err != nil {
-		return nil, 0, err
-	}
-	d := time.Duration(www.VerificationExpiryHours) * time.Hour
-	expiry := time.Now().Add(d).Unix()
-	return tokenb, expiry, nil
-}
-
-// initUserEmailsCache initializes the userEmails cache by iterating through
-// all the users in the database and adding a email-userID mapping for them.
-//
-// This function must be called WITHOUT the lock held.
-func (p *politeiawww) initUserEmailsCache() error {
-	p.Lock()
-	defer p.Unlock()
-
-	return p.db.AllUsers(func(u *user.User) {
-		p.userEmails[u.Email] = u.ID
-	})
-}
-
-// setUserEmailsCache sets a email-userID mapping in the user emails cache.
-//
-// This function must be called WITHOUT the lock held.
-func (p *politeiawww) setUserEmailsCache(email string, id uuid.UUID) {
-	p.Lock()
-	defer p.Unlock()
-	p.userEmails[email] = id
-}
-
-// userIDByEmail returns a userID given their email address.
-//
-// This function must be called WITHOUT the lock held.
-func (p *politeiawww) userIDByEmail(email string) (uuid.UUID, bool) {
-	p.RLock()
-	defer p.RUnlock()
-	id, ok := p.userEmails[email]
-	return id, ok
-}
-
-// userByEmail returns a User object given their email address.
-//
-// This function must be called WITHOUT the lock held.
-func (p *politeiawww) userByEmail(email string) (*user.User, error) {
-	id, ok := p.userIDByEmail(email)
-	if !ok {
-		log.Debugf("userByEmail: email lookup failed for '%v'", email)
-		return nil, user.ErrUserNotFound
-	}
-	return p.db.UserGetById(id)
-}
-
-// userByIDStr converts the provided userIDStr to a uuid and returns the
-// corresponding user, if one it exists.
-func (p *politeiawww) userByIDStr(userIDStr string) (*user.User, error) {
-	userID, err := uuid.Parse(userIDStr)
-	if err != nil {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusInvalidUUID,
-		}
-	}
-
-	usr, err := p.db.UserGetById(userID)
-	if err != nil {
-		if errors.Is(err, user.ErrUserNotFound) {
-			return nil, www.UserError{
-				ErrorCode: www.ErrorStatusUserNotFound,
-			}
-		}
-		return nil, err
-	}
-
-	return usr, nil
-}
-
-// hashPassword hashes the given password string with the default bcrypt cost
-// or the minimum cost if the test flag is set to speed up running tests.
-func (p *politeiawww) hashPassword(password string) ([]byte, error) {
-	if p.test {
-		return bcrypt.GenerateFromPassword([]byte(password),
-			bcrypt.MinCost)
-	}
-	return bcrypt.GenerateFromPassword([]byte(password),
-		bcrypt.DefaultCost)
-}
-
-// processNewUser creates a new user in the db if it doesn't already
-// exist and sets a verification token and expiry; the token must be
-// verified before it expires. If the user already exists in the db
-// and its token is expired, it generates a new one.
-//
-// Note that this function always returns a NewUserReply. The caller shall
-// verify error and determine how to return this information upstream.
-func (p *politeiawww) processNewUser(nu www.NewUser) (*www.NewUserReply, error) {
-	log.Tracef("processNewUser: %v", nu.Username)
-
-	// Format and validate user credentials
-	nu.Email = strings.ToLower(nu.Email)
-	if !validEmail.MatchString(nu.Email) {
-		log.Debugf("processNewUser: invalid email '%v'", nu.Email)
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusMalformedEmail,
-		}
-	}
-	err := validatePubKey(nu.PublicKey)
-	if err != nil {
-		return nil, err
-	}
-
-	nu.Username = formatUsername(nu.Username)
-	err = validateUsername(nu.Username)
-	if err != nil {
-		return nil, err
-	}
-
-	// The client should be hashing the password before
-	// sending it to politeiawww. This validation is only
-	// relevant if the client failed to hash the password
-	// or does not include a password in the request.
-	err = validatePassword(nu.Password)
-	if err != nil {
-		return nil, err
-	}
-
-	// Check if user already exists
-	u, err := p.userByEmail(nu.Email)
-	switch err {
-	case nil:
-		// User exists
-
-		// Return if the user is already verified
-		if u.NewUserVerificationToken == nil {
-			log.Debugf("processNewUser: '%v' already verified",
-				nu.Username)
-			return &www.NewUserReply{}, nil
-		}
-
-		// User is not already verified. Check if the verification token
-		// has expired. If the token has not expired yet, we simply return.
-		// The user will have to wait until the token expires before a
-		// new one will be sent. If the token has expired, we update the
-		// user in the database and send the user a new token. The user
-		// identity will be updated if the request specifies a new public
-		// key.
-		if u.NewUserVerificationExpiry > time.Now().Unix() {
-			log.Debugf("processNewUser: '%v' not verified, "+
-				"token not expired", nu.Username)
-			return &www.NewUserReply{}, nil
-		}
-
-		// Ensure public key is unique
-		usr, err := p.db.UserGetByPubKey(nu.PublicKey)
-		if err != nil {
-			if errors.Is(err, user.ErrUserNotFound) {
-				// Pubkey is unique, but is not the same pubkey that
-				// the user originally signed up with. This is fine.
-				// The user's identity just needs to be updated.
-				id, err := user.NewIdentity(nu.PublicKey)
-				if err != nil {
-					return nil, err
-				}
-				err = u.AddIdentity(*id)
-				if err != nil {
-					return nil, err
-				}
-			}
-		} else {
-			switch {
-			case usr.ID.String() == u.ID.String():
-				// Pubkey exists and belongs to this user. This is
-				// ok. It just means that the user is requesting a
-				// new verification token using the same identity
-				// that they signed up with. Continue.
-			default:
-				// Pubkey exists and belongs to another user
-				return nil, www.UserError{
-					ErrorCode: www.ErrorStatusDuplicatePublicKey,
-				}
-			}
-		}
-
-		// Generate a new verification token
-		tokenb, expiry, err := newVerificationTokenAndExpiry()
-		if err != nil {
-			return nil, err
-		}
-
-		// Email the verification token before updating the
-		// database. If the email fails, the database won't
-		// be updated.
-		err = p.emailUserEmailVerify(u.Email,
-			hex.EncodeToString(tokenb), u.Username)
-		if err != nil {
-			log.Errorf("processNewUser: mail verification "+
-				"token failed for '%v': %v", u.Email, err)
-			return &www.NewUserReply{}, nil
-		}
-
-		// Update user record with the verification token and
-		// the new identity if one was set.
-		u.NewUserVerificationToken = tokenb
-		u.NewUserVerificationExpiry = expiry
-		err = p.db.UserUpdate(*u)
-		if err != nil {
-			return nil, err
-		}
-
-		// Send reply. Only return the verification token in
-		// the reply if the mail server has been disabled.
-		var t string
-		if !p.mail.IsEnabled() {
-			t = hex.EncodeToString(u.NewUserVerificationToken)
-		}
-		return &www.NewUserReply{
-			VerificationToken: t,
-		}, nil
-	case user.ErrUserNotFound:
-		// User doesn't exist; continue
-	default:
-		// All other errors
-		return nil, err
-	}
-
-	// User does not exist. Create a new user.
-
-	// Ensure username is unique
-	_, err = p.db.UserGetByUsername(nu.Username)
-	switch err {
-	case nil:
-		// Duplicate username
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusDuplicateUsername,
-		}
-	case user.ErrUserNotFound:
-		// Username does not exist; continue
-	default:
-		return nil, err
-	}
-
-	// Ensure public key is unique
-	_, err = p.db.UserGetByPubKey(nu.PublicKey)
-	switch err {
-	case user.ErrUserNotFound:
-		// Pubkey is unique; continue
-	case nil:
-		// Duplicate pubkey
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusDuplicatePublicKey,
-		}
-	default:
-		// All other errors
-		return nil, err
-	}
-
-	// Create user
-	hashedPass, err := p.hashPassword(nu.Password)
-	if err != nil {
-		return nil, err
-	}
-	tokenb, expiry, err := newVerificationTokenAndExpiry()
-	if err != nil {
-		return nil, err
-	}
-	newUser := user.User{
-		Email:                     nu.Email,
-		Username:                  nu.Username,
-		HashedPassword:            hashedPass,
-		NewUserVerificationToken:  tokenb,
-		NewUserVerificationExpiry: expiry,
-	}
-	id, err := user.NewIdentity(nu.PublicKey)
-	if err != nil {
-		return nil, err
-	}
-	err = newUser.AddIdentity(*id)
-	if err != nil {
-		return nil, err
-	}
-
-	// Try to email the verification link first; if it fails,
-	// then the new user won't be created.
-	//
-	// This is conditional on the email server being setup.
-	err = p.emailUserEmailVerify(newUser.Email,
-		hex.EncodeToString(tokenb), newUser.Username)
-	if err != nil {
-		log.Errorf("processNewUser: mail verification token "+
-			"failed for '%v': %v", newUser.Email, err)
-		return &www.NewUserReply{}, nil
-	}
-
-	// Save new user to the database
-	err = p.db.UserNew(newUser)
-	if err != nil {
-		return nil, err
-	}
-
-	// Set paywall info for the user. This had to wait until after the
-	// user was already created in the db because the db sets the
-	// paywall address index when the user is created, and the paywall
-	// address index is used to generate the paywall info. Lookup the
-	// user from the db to get the paywall address index.
-	u, err = p.db.UserGetByUsername(newUser.Username)
-	if err != nil {
-		return nil, err
-	}
-	err = p.GenerateNewUserPaywall(u)
-	if err != nil {
-		return nil, err
-	}
-
-	// Update memory cache
-	p.setUserEmailsCache(u.Email, u.ID)
-
-	log.Infof("New user created: %v", u.Username)
-
-	// Only return the verification token in the reply
-	// if the mail server has been disabled.
-	var t string
-	if !p.mail.IsEnabled() {
-		t = hex.EncodeToString(u.NewUserVerificationToken)
-	}
-	return &www.NewUserReply{
-		VerificationToken: t,
-	}, nil
-}
-
 // processUserDetails return the requested user's details. Some fields can be
 // omitted or blank depending on the requester's access level.
 func (p *politeiawww) processUserDetails(ud *www.UserDetails, isCurrentUser bool, isAdmin bool) (*www.UserDetailsReply, error) {
@@ -633,14 +134,6 @@ func (p *politeiawww) processEditUser(eu *www.EditUser, user *user.User) (*www.E
 	}
 
 	return &www.EditUserReply{}, nil
-}
-
-// userHasPaid returns whether the user has paid the user registration paywall.
-func (p *politeiawww) userHasPaid(u user.User) bool {
-	if !p.paywallIsEnabled() {
-		return true
-	}
-	return u.NewUserPaywallTx != ""
 }
 
 // createLoginReply creates a login reply.
@@ -2059,256 +1552,6 @@ func (p *politeiawww) processUserPaymentsRescan(ctx context.Context, upr www.Use
 	return &www.UserPaymentsRescanReply{
 		NewCredits: newCreditsWWW,
 	}, nil
-}
-
-// removeUsersFromPool removes the provided user IDs from the the poll pool.
-//
-// Currently, updating the user db and removing the user from pool isn't an
-// atomic operation.  This can lead to a scenario where the user has been
-// marked as paid in the db, but has not yet been removed from the pool. If a
-// user issues a proposal paywall during this time, the proposal paywall will
-// replace the user paywall in the pool. When the pool proceeds to remove the
-// user paywall, it will mistakenly remove the proposal paywall instead.
-// Proposal credits will not be added to the user's account. The workaround
-// until this code gets replaced with websockets is to pass in the paywallType
-// when removing a pool member.
-//
-// This function must be called WITHOUT the mutex held.
-func (p *politeiawww) removeUsersFromPool(userIDsToRemove []uuid.UUID, paywallType string) {
-	p.Lock()
-	defer p.Unlock()
-
-	for _, userID := range userIDsToRemove {
-		if p.userPaywallPool[userID].paywallType == paywallType {
-			delete(p.userPaywallPool, userID)
-		}
-	}
-}
-
-// addUserToPaywallPool adds a database user to the paywall pool.
-//
-// This function must be called WITH the mutex held.
-func (p *politeiawww) addUserToPaywallPool(u *user.User, paywallType string) {
-	p.userPaywallPool[u.ID] = paywallPoolMember{
-		paywallType: paywallType,
-		address:     u.NewUserPaywallAddress,
-		amount:      u.NewUserPaywallAmount,
-		txNotBefore: u.NewUserPaywallTxNotBefore,
-		pollExpiry:  u.NewUserPaywallPollExpiry,
-	}
-}
-
-// addUserToPaywallPoolLock adds a user and its paywall info to the in-memory pool.
-//
-// This function must be called WITHOUT the mutex held.
-func (p *politeiawww) addUserToPaywallPoolLock(u *user.User, paywallType string) {
-	if !p.paywallIsEnabled() {
-		return
-	}
-
-	p.Lock()
-	defer p.Unlock()
-
-	p.addUserToPaywallPool(u, paywallType)
-}
-
-// addUsersToPaywallPool adds a user and its paywall info to the in-memory pool.
-//
-// This function must be called WITHOUT the mutex held.
-func (p *politeiawww) addUsersToPaywallPool() error {
-	p.Lock()
-	defer p.Unlock()
-
-	// Create the in-memory pool of all users who need to pay the paywall.
-	err := p.db.AllUsers(func(u *user.User) {
-		// Proposal paywalls
-		if p.userHasValidProposalPaywall(u) {
-			p.addUserToPaywallPool(u, paywallTypeProposal)
-			return
-		}
-
-		// User paywalls
-		if p.userHasPaid(*u) {
-			return
-		}
-		if u.NewUserVerificationToken != nil {
-			return
-		}
-		if paywallHasExpired(u.NewUserPaywallPollExpiry) {
-			return
-		}
-
-		p.addUserToPaywallPool(u, paywallTypeUser)
-	})
-	if err != nil {
-		return err
-	}
-
-	log.Tracef("Adding %v users to paywall pool", len(p.userPaywallPool))
-	return nil
-}
-
-// updateUserAsPaid records in the database that the user has paid.
-func (p *politeiawww) updateUserAsPaid(u *user.User, tx string) error {
-	u.NewUserPaywallTx = tx
-	u.NewUserPaywallPollExpiry = 0
-	return p.db.UserUpdate(*u)
-}
-
-// derivePaywallInfo derives a new paywall address for the user.
-func (p *politeiawww) derivePaywallInfo(u *user.User) (string, uint64, int64, error) {
-	address, err := util.DerivePaywallAddress(p.params,
-		p.cfg.PaywallXpub, uint32(u.PaywallAddressIndex))
-	if err != nil {
-		err = fmt.Errorf("Unable to derive paywall address #%v "+
-			"for %v: %v", u.ID.ID(), u.Email, err)
-	}
-
-	return address, p.cfg.PaywallAmount, time.Now().Unix(), err
-}
-
-// createUserPaywallPoolCopy returns a map of the poll pool.
-//
-// This function must be called WITHOUT the mutex held.
-func (p *politeiawww) createUserPaywallPoolCopy() map[uuid.UUID]paywallPoolMember {
-	p.RLock()
-	defer p.RUnlock()
-
-	poolCopy := make(map[uuid.UUID]paywallPoolMember, len(p.userPaywallPool))
-
-	for k, v := range p.userPaywallPool {
-		poolCopy[k] = v
-	}
-
-	return poolCopy
-}
-
-// checkForUserPayments is called periodically to see if payments have come
-// through.
-func (p *politeiawww) checkForUserPayments(ctx context.Context, pool map[uuid.UUID]paywallPoolMember) (bool, []uuid.UUID) {
-	var userIDsToRemove []uuid.UUID
-
-	for userID, poolMember := range pool {
-		u, err := p.db.UserGetById(userID)
-		if err != nil {
-			if errors.Is(err, user.ErrShutdown) {
-				// The database is shutdown, so stop the
-				// thread.
-				return false, nil
-			}
-
-			log.Errorf("cannot fetch user by id %v: %v\n",
-				userID, err)
-			continue
-		}
-
-		if poolMember.paywallType != paywallTypeUser {
-			continue
-		}
-
-		log.Tracef("Checking the user paywall address for user %v...",
-			u.Email)
-
-		if p.userHasPaid(*u) {
-			// The user could have been marked as paid by
-			// RouteUserRegistrationPayment, so just remove him from the
-			// in-memory pool.
-			userIDsToRemove = append(userIDsToRemove, userID)
-			log.Tracef("  removing from polling, user already paid")
-			continue
-		}
-
-		if paywallHasExpired(u.NewUserPaywallPollExpiry) {
-			userIDsToRemove = append(userIDsToRemove, userID)
-			log.Tracef("  removing from polling, poll has expired")
-			continue
-		}
-
-		tx, _, err := util.FetchTxWithBlockExplorers(ctx, p.params, poolMember.address,
-			poolMember.amount, poolMember.txNotBefore,
-			p.cfg.MinConfirmationsRequired, p.dcrdataHostHTTP())
-		if err != nil {
-			log.Errorf("cannot fetch tx: %v\n", err)
-			continue
-		}
-
-		if tx != "" {
-			// Update the user in the database.
-			err = p.updateUserAsPaid(u, tx)
-			if err != nil {
-				if errors.Is(err, user.ErrShutdown) {
-					// The database is shutdown, so stop
-					// the thread.
-					return false, nil
-				}
-
-				log.Errorf("cannot update user with id %v: %v",
-					u.ID, err)
-				continue
-			}
-
-			// Remove this user from the in-memory pool.
-			userIDsToRemove = append(userIDsToRemove, userID)
-			log.Tracef("  removing from polling, user just paid")
-		}
-
-		time.Sleep(paywallCheckGap)
-	}
-
-	return true, userIDsToRemove
-}
-
-// GenerateNewUserPaywall generates new paywall info, if necessary, and saves
-// it in the database.
-func (p *politeiawww) GenerateNewUserPaywall(u *user.User) error {
-	// Check that the paywall is enabled.
-	if !p.paywallIsEnabled() {
-		return nil
-	}
-
-	// Check that the user either hasn't had paywall information set yet,
-	// or it has expired.
-	if u.NewUserPaywallAddress != "" &&
-		!paywallHasExpired(u.NewUserPaywallPollExpiry) {
-		return nil
-	}
-
-	if u.NewUserPaywallAddress == "" {
-		address, amount, txNotBefore, err := p.derivePaywallInfo(u)
-		if err != nil {
-			return err
-		}
-
-		u.NewUserPaywallAddress = address
-		u.NewUserPaywallAmount = amount
-		u.NewUserPaywallTxNotBefore = txNotBefore
-	}
-	u.NewUserPaywallPollExpiry = time.Now().Add(paywallExpiryDuration).Unix()
-
-	err := p.db.UserUpdate(*u)
-	if err != nil {
-		return err
-	}
-
-	p.addUserToPaywallPoolLock(u, paywallTypeUser)
-	return nil
-}
-
-// initPaywallCheck is intended to be called
-func (p *politeiawww) initPaywallChecker() error {
-	if p.cfg.PaywallAmount == 0 {
-		// Paywall not configured.
-		return nil
-	}
-
-	err := p.addUsersToPaywallPool()
-	if err != nil {
-		return err
-	}
-
-	// Start the thread that checks for payments.
-	go p.checkForPayments()
-	return nil
 }
 
 func (p *politeiawww) totpCheck(code string, u *user.User) error {
