+++ politeiawww/www.go
@@ -19,7 +19,6 @@ import (
 	"os"
 	"os/signal"
 	"path/filepath"
-	"runtime/debug"
 	"strings"
 	"syscall"
 	"time"
@@ -28,8 +27,6 @@ import (
 	pd "github.com/decred/politeia/politeiad/api/v1"
 	pdv2 "github.com/decred/politeia/politeiad/api/v2"
 	pdclient "github.com/decred/politeia/politeiad/client"
-	cms "github.com/decred/politeia/politeiawww/api/cms/v1"
-	www "github.com/decred/politeia/politeiawww/api/www/v1"
 	database "github.com/decred/politeia/politeiawww/cmsdatabase"
 	cmsdb "github.com/decred/politeia/politeiawww/cmsdatabase/cockroachdb"
 	ghtracker "github.com/decred/politeia/politeiawww/codetracker/github"
@@ -59,42 +56,6 @@ const (
 
 	csrfKeyLength = 32
 )
-
-func convertWWWErrorStatusFromPD(e pd.ErrorStatusT) www.ErrorStatusT {
-	switch e {
-	case pd.ErrorStatusInvalidRequestPayload:
-		// Intentionally omitted because this indicates a politeiawww
-		// server error so a ErrorStatusInvalid should be returned.
-	case pd.ErrorStatusInvalidChallenge:
-		// Intentionally omitted because this indicates a politeiawww
-		// server error so a ErrorStatusInvalid should be returned.
-	case pd.ErrorStatusInvalidFilename:
-		return www.ErrorStatusInvalidFilename
-	case pd.ErrorStatusInvalidFileDigest:
-		return www.ErrorStatusInvalidFileDigest
-	case pd.ErrorStatusInvalidBase64:
-		return www.ErrorStatusInvalidBase64
-	case pd.ErrorStatusInvalidMIMEType:
-		return www.ErrorStatusInvalidMIMEType
-	case pd.ErrorStatusUnsupportedMIMEType:
-		return www.ErrorStatusUnsupportedMIMEType
-	case pd.ErrorStatusInvalidRecordStatusTransition:
-		return www.ErrorStatusInvalidPropStatusTransition
-	}
-	return www.ErrorStatusInvalid
-}
-
-func convertWWWErrorStatus(pluginID string, errCode int) www.ErrorStatusT {
-	switch pluginID {
-	case "":
-		// politeiad API
-		e := pd.ErrorStatusT(errCode)
-		return convertWWWErrorStatusFromPD(e)
-	}
-
-	// No corresponding www error status found
-	return www.ErrorStatusInvalid
-}
 
 // Fetch remote identity
 func getIdentity(rpcHost, rpcCert, rpcIdentityFile, interactive string) error {
@@ -133,134 +94,6 @@ func getIdentity(rpcHost, rpcCert, rpcIdentityFile, interactive string) error {
 	log.Infof("Identity saved to: %v", rpcIdentityFile)
 
 	return nil
-}
-
-// userErrorStatus retrieves the human readable error message for an error
-// status code. The status code can be from either the pi or cms api.
-func userErrorStatus(e www.ErrorStatusT) string {
-	s, ok := www.ErrorStatus[e]
-	if ok {
-		return s
-	}
-	s, ok = cms.ErrorStatus[e]
-	if ok {
-		return s
-	}
-	return ""
-}
-
-// RespondWithError returns an HTTP error status to the client. If it's a user
-// error, it returns a 4xx HTTP status and the specific user error code. If it's
-// an internal server error, it returns 500 and an error code which is also
-// outputted to the logs so that it can be correlated later if the user
-// files a complaint.
-func RespondWithError(w http.ResponseWriter, r *http.Request, userHttpCode int, format string, args ...interface{}) {
-	// XXX this function needs to get an error in and a format + args
-	// instead of what it is doing now.
-	// So inError error, format string, args ...interface{}
-	// if err == nil -> internal error using format + args
-	// if err != nil -> if defined error -> return defined error + log.Errorf format+args
-	// if err != nil -> if !defined error -> return + log.Errorf format+args
-
-	// Check if the client dropped the connection
-	if err := r.Context().Err(); err == context.Canceled {
-		log.Infof("%v %v %v %v client aborted connection",
-			util.RemoteAddr(r), r.Method, r.URL, r.Proto)
-
-		// Client dropped the connection. There is no need to
-		// respond further.
-		return
-	}
-
-	// Check for www user error
-	if userErr, ok := args[0].(www.UserError); ok {
-		// Error is a www user error. Log it and return a 400.
-		if userHttpCode == 0 {
-			userHttpCode = http.StatusBadRequest
-		}
-
-		if len(userErr.ErrorContext) == 0 {
-			log.Infof("WWW user error: %v %v %v",
-				util.RemoteAddr(r), int64(userErr.ErrorCode),
-				userErrorStatus(userErr.ErrorCode))
-		} else {
-			log.Infof("WWW user error: %v %v %v: %v",
-				util.RemoteAddr(r), int64(userErr.ErrorCode),
-				userErrorStatus(userErr.ErrorCode),
-				strings.Join(userErr.ErrorContext, ", "))
-		}
-
-		util.RespondWithJSON(w, userHttpCode,
-			www.UserError{
-				ErrorCode:    userErr.ErrorCode,
-				ErrorContext: userErr.ErrorContext,
-			})
-		return
-	}
-
-	// Check for politeiad error
-	if pdError, ok := args[0].(pdError); ok {
-		var (
-			pluginID   = pdError.ErrorReply.Plugin
-			errCode    = pdError.ErrorReply.ErrorCode
-			errContext = pdError.ErrorReply.ErrorContext
-		)
-
-		// Check if the politeiad error corresponds to a www user error
-		wwwErrCode := convertWWWErrorStatus(pluginID, errCode)
-		if wwwErrCode == www.ErrorStatusInvalid {
-			// politeiad error does not correspond to a www user error. Log
-			// it and return a 500.
-			t := time.Now().Unix()
-			if pluginID == "" {
-				log.Errorf("%v %v %v %v Internal error %v: error "+
-					"code from politeiad: %v", util.RemoteAddr(r), r.Method,
-					r.URL, r.Proto, t, errCode)
-			} else {
-				log.Errorf("%v %v %v %v Internal error %v: error "+
-					"code from politeiad plugin %v: %v", util.RemoteAddr(r),
-					r.Method, r.URL, r.Proto, t, pluginID, errCode)
-			}
-
-			util.RespondWithJSON(w, http.StatusInternalServerError,
-				www.ErrorReply{
-					ErrorCode: t,
-				})
-			return
-		}
-
-		// politeiad error does correspond to a www user error. Log it
-		// and return a 400.
-		if len(errContext) == 0 {
-			log.Infof("WWW user error: %v %v %v",
-				util.RemoteAddr(r), int64(wwwErrCode),
-				userErrorStatus(wwwErrCode))
-		} else {
-			log.Infof("WWW user error: %v %v %v: %v",
-				util.RemoteAddr(r), int64(wwwErrCode),
-				userErrorStatus(wwwErrCode),
-				strings.Join(errContext, ", "))
-		}
-
-		util.RespondWithJSON(w, http.StatusBadRequest,
-			www.UserError{
-				ErrorCode:    wwwErrCode,
-				ErrorContext: errContext,
-			})
-		return
-	}
-
-	// Error is a politeiawww server error. Log it and return a 500.
-	t := time.Now().Unix()
-	ec := fmt.Sprintf("%v %v %v %v Internal error %v: ", util.RemoteAddr(r),
-		r.Method, r.URL, r.Proto, t)
-	log.Errorf(ec+format, args...)
-	log.Errorf("Stacktrace (NOT A REAL CRASH): %s", debug.Stack())
-
-	util.RespondWithJSON(w, http.StatusInternalServerError,
-		www.ErrorReply{
-			ErrorCode: t,
-		})
 }
 
 // addRoute sets up a handler for a specific method+route. If method is not
